requirements-apple-m2.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile setup.cfg
#
absl-py==1.1.0
    # via mediapipe
attrs==21.4.0
    # via mediapipe
certifi==2023.7.22
    # via requests
chardet==4.0.0
    # via requests
click==8.0.2
    # via dot (setup.cfg)
cycler==0.11.0
    # via matplotlib
dlib==19.19.0
    # via dot (setup.cfg)
face-alignment==1.3.3
    # via dot (setup.cfg)
flatbuffers==2.0
    # via onnxruntime
fonttools==4.43.0
    # via matplotlib
idna==2.10
    # via requests
imageio==2.19.3
    # via scikit-image
kiwisolver==1.4.3
    # via matplotlib
kornia==0.6.5
    # via dot (setup.cfg)
llvmlite==0.38.1
    # via numba
matplotlib==3.5.2
    # via mediapipe
mediapipe-silicon
    # via dot (setup.cfg)
mediapipe==0.10.3
networkx==2.8.4
    # via scikit-image
numba==0.55.2
    # via face-alignment
numpy==1.22.0
    # via
    #   dot (setup.cfg)
    #   face-alignment
    #   imageio
    #   matplotlib
    #   mediapipe
    #   numba
    #   onnxruntime
    #   opencv-contrib-python
    #   opencv-python
    #   pywavelets
    #   scikit-image
    #   scipy
    #   tifffile
    #   torchvision
onnxruntime==1.15.1
    # via dot (setup.cfg)
opencv-contrib-python==4.5.5.62
    # via
    #   dot (setup.cfg)
    #   mediapipe
opencv-python==4.5.5.62
    # via
    #   dot (setup.cfg)
    #   face-alignment
packaging==21.3
    # via
    #   kornia
    #   matplotlib
    #   scikit-image
pillow==10.0.1
    # via
    #   dot (setup.cfg)
    #   imageio
    #   matplotlib
    #   scikit-image
    #   torchvision
protobuf==3.20.2
    # via
    #   dot (setup.cfg)
    #   mediapipe
    #   onnxruntime
pyparsing==3.0.9
    # via
    #   matplotlib
    #   packaging
python-dateutil==2.8.2
    # via matplotlib
pywavelets==1.3.0
    # via scikit-image
pyyaml==5.4.1
    # via dot (setup.cfg)
requests==2.31.0
    # via dot (setup.cfg)
scikit-image==0.19.1
    # via
    #   dot (setup.cfg)
    #   face-alignment
scipy==1.10.1
    # via
    #   dot (setup.cfg)
    #   face-alignment
    #   scikit-image
six==1.16.0
    # via
    #   mediapipe
    #   python-dateutil
tifffile==2022.5.4
    # via scikit-image
torch==2.0.1
    # via
    #   dot (setup.cfg)
    #   face-alignment
    #   kornia
    #   torchvision
torchvision==0.15.2
    # via dot (setup.cfg)
tqdm==4.64.0
    # via face-alignment
typing-extensions==4.3.0
    # via torch
urllib3==1.26.18
    # via requests
wheel==0.38.1
    # via mediapipe

# The following packages are considered to be unsafe in a requirements file:
# setuptools
